requirements.txt
# !! use python=3.10.*
numpy==1.25
diffusers==0.24.0
huggingface-hub==0.25.2
torch==2.1.0
controlnet-aux==0.0.7
transformers==4.35.2
xformers==0.0.21
accelerate==0.25.0
triton==2.0.0
Pillow==11.1.0
opencv-python==4.11.0.86
tqdm==4.67.1
spandrel==0.4.1
boto3==1.37.4
python-dotenv==1.0.1
pymongo==4.11.1
fastapi[standard]